(GH-271) Move classes into extensions folder
To be consistent with where another extension class lives
